+ refactor Slave to standalone puppet.js class for better unit test and naming
