+++ lib/master.js
@@ -9,6 +9,7 @@ var _ = require('underscore'),
 	fork = require('child_process').fork,
 	BigNumber = require('bignumber.js'),
 	Worker = require('./worker.js').Worker,
+	Puppet = require('./puppet.js').Puppet,
 	rejectIfPortBusy = require('./utils.js').rejectIfPortBusy,
 	writePid = require('./utils.js').writePid,
 	getNodeInspectorPath = require('./utils.js').getNodeInspectorPath,
@@ -16,194 +17,6 @@ var _ = require('underscore'),
 
 var masterDeferred = when.defer(),
 	masterPromise = exports.master = masterDeferred.promise;
-
-/**
- * Slave is an image of the worker at the Master's runtime, to simplify the state machine
- */
-var Slave = exports.Slave = function Slave(master, worker, options, env){
-
-	var _this = this;
-
-	_.extend(_this, {
-		'logger': master.logger,
-		'emitter': master.emitter,
-		'cacheManager': env.CACHE_MANAGER,
-		'debugging': env.debug,
-		'worker': worker,
-		'pid': worker.process.pid
-	});
-
-	_this.forkedState = {
-
-		'disconnect': function(){
-			
-			worker.disconnect();
-		},
-
-		'whenOnline': function(){
-
-			if(_this.debugging){
-				//we could enable debugging here
-				master.emitter.once('debug-started', function(){
-					_this.emitter.emit('run', [_this.pid]);
-				});
-
-				master.emitter.once('debug-finished', function(){
-					_this.disconnect();
-				});
-
-				_this.debug();
-			}
-		},
-
-		'whenListening': function(){
-
-			_this.state = _this.activeState;
-			//this is mainly for the old worker to be notified that the successor is ready
-			master.emitter.emit(util.format('worker-%d-listening', worker.process.pid), ['self']);
-		},
-
-		'whenHeartbeat': function(){
-			//no interest
-		},
-
-		'whenExit': function(){
-
-			if(!worker.suicide){//accident, must be revived
-				master.fork(options, env);
-			}
-		}
-	};
-
-	_this.activeState = {
-
-		'disconnect': function(){
-			
-			worker.disconnect();
-			worker.suicide = true;
-		},
-
-		'whenHeartbeat': function(heartbeat){
-			//here's the key feature for cluster3, based on the historic tps info, memory usage, gc rate, we could determine if a slave should
-			//enter an old state from active
-
-			if(master.markOld(heartbeat)){
-
-				var dieDeferred = when.defer(),
-					diePromise = null,
-					die = function(){
-
-						_this.activeState = _this.oldState;
-
-						var successor = master.fork(options, env);
-						//when successor is in place, the old worker could be discontinued finally
-						master.emitter.once(util.format('worker-%d-listening', successor.process.pid), function(){
-
-							master.emitter.emit('disconnect', ['master', worker.pid]);
-							dieDeferred.resolve(worker.pid);
-
-							if(master.dyingQueue === diePromise){//last of dyingQueue resolved, clean up the dyingQueue
-								master.dyingQueue = null;
-							}
-						});
-					};
-
-				if(!master.dyingQueue){//1st in the dying queue, 
-					diePromise = master.dyingQueue = timeout(dieDeferred.promise, 60000);//1 min
-					die();
-				}
-				else{
-					diePromise = master.dyingQueue = timeout(master.dyingQueue, 60000).ensure(die);
-				}
-			}
-		},
-
-		'whenExit': function(){
-
-			if(!worker.suicide){//accident, must be revived
-				master.fork(options, env);
-			}
-			//already exit, clean up
-			delete master.slaves[_this.pid];
-		}
-	};
-
-	_this.oldState = {
-
-		'disconnect': function(){
-			
-			worker.disconnect();
-		},
-
-		'whenExit': function(){
-
-			_this.state = _this.diedState;
-
-			//already exit, clean up
-			delete master.slaves[_this.pid];
-		}
-	};
-
-	_this.diedState = {
-
-		'disconnect': function(){
-			//already disconnected	
-		},
-
-		'whenExit': function(){
-			//already exit, clean up
-			delete master.slaves[_this.pid];
-		}
-	};
-
-	_this.state = _this.forkedState;
-};
-
-Slave.prototype.debug = function(){
-
-	//debugger listening on port 5858
-	var _this = this,
-		buff = '',
-		pid = _this.pid,
-		worker = _this.worker,
-		debugDeferred = when.defer();
-
-	_this.emitter.once('debug-port-listening', function(){
-
-		_this.logger.info('[debug] slave: %d received debug-port-listening', pid);
-		debugDeferred.resolve(pid);
-	});
-
-	worker.process.kill('SIGUSR1');//make it debug ready
-	_this.logger.debug('[debug] master sent signal SIGUSR1 to %d', pid);
-
-	return debugDeferred.promise;
-};
-
-Slave.prototype.disconnect = function(){
-
-	return this.state.disconnect();
-};
-
-Slave.prototype.whenOnline = function(){
-
-	return this.state.whenOnline();
-};
-
-Slave.prototype.whenListening = function(){
-
-	return this.state.whenListening();
-};
-
-Slave.prototype.whenHeartbeat = function(heartbeat){
-
-	return this.state.whenHeartbeat(heartbeat);
-};
-
-Slave.prototype.whenExit = function(){
-
-	return this.state.whenExit();
-};
 
 var Master = exports.Master = function(proc, options){
 
@@ -242,7 +55,7 @@ Master.prototype.initialize = function(proc, options){
 					return _this.fork(_this.options);
 				});
 
-				//wait for all slaves to be ready, and then enable ECV as the last step
+				//wait for all puppets to be ready, and then enable ECV as the last step
 				_this.markUp();	
 			},
 			'noWorkers': options.noWorkers,
@@ -261,20 +74,20 @@ Master.prototype.initialize = function(proc, options){
 				
 				var _this = this,
 					pid = heartbeat.pid,
-					slave = _this.slaves[pid],
+					puppet = _this.puppets[pid],
 					currTPS = heartbeat.durations / heartbeat.transactions;
 
 				if(currTPS <= 2){//TPS too low, no good for sampling as the 1st phase.
 					return false;
 				}
 
-				var peak = slave.peak = slave.peak || {
+				var peak = puppet.peak = puppet.peak || {
 						'tps': currTPS,
 						'cpu': heartbeat.cpu,
 						'memory': heartbeat.memory,
 						'gc.incremental': heartbeat.gc.incremental,
 						'gc.full': heartbeat.gc.full
-					};//remember the peak of each slave
+					};//remember the peak of each puppet
 
 				if(currTPS >= peak.tps){
 					peak.tps = Math.max(currTPS.tps, peak.tps);
@@ -300,32 +113,32 @@ Master.prototype.initialize = function(proc, options){
 			'debug.saveLiveEdit': options.debug.saveLiveEdit,
 			'debug.hidden': options.debug.hidden || [],
 			'options': options,
-			'slaves': {
+			'puppets': {
 
 			},
 			'status': require('./status')
 		});
 
 		cluster.on('fork', function(worker){//interestingly, on fork or #fork could complete in disorder...
-			_this.slaves[worker.process.pid] = _this.slaves[worker.process.pid] || new Slave(_this, worker, options, worker.process.env || {});
-			_this.slaves[worker.process.pid].worker = worker;
+			_this.puppets[worker.process.pid] = _this.puppets[worker.process.pid] || new Puppet(_this, worker, options, worker.process.env || {});
+			_this.puppets[worker.process.pid].worker = worker;
 		});
 
 		cluster.on('online', function(worker){
-			_this.slaves[worker.process.pid].whenOnline();
+			_this.puppets[worker.process.pid].whenOnline();
 		});
 
 		cluster.on('listening', function(worker){
-			_this.slaves[worker.process.pid].whenListening();
+			_this.puppets[worker.process.pid].whenListening();
 		});
 
 		cluster.on('exit', function(worker){
 			_this.logger.info('master detects exit of worker:' + worker.process.pid);
-			_this.slaves[worker.process.pid].whenExit();
+			_this.puppets[worker.process.pid].whenExit();
 		});
 
 		_this.emitter.on('heartbeat', function(heartbeat){
-			_this.slaves[heartbeat.pid].whenHeartbeat(heartbeat);
+			_this.puppets[heartbeat.pid].whenHeartbeat(heartbeat);
 		});
 
 		process.once('SIGINT', _.bind(_this.whenStop, _this));
@@ -432,15 +245,15 @@ Master.prototype.debug = function debug(pid){
 	_this.markDown();
 
 	//debug fresh
-	_.each(_this.slaves, function(slave){
+	_.each(_this.puppets, function(puppet){
 
-		if(!slave.cacheManager && pid !== slave.pid){
-			_this.logger.info('debug fresh, disconnecting:%j vs pid:%j', slave.pid, pid);
-			slave.disconnect();
+		if(!puppet.cacheManager && pid !== puppet.pid){
+			_this.logger.info('debug fresh, disconnecting:%j vs pid:%j', puppet.pid, pid);
+			puppet.disconnect();
 		}
 	});
 
-	//send signal to the slave
+	//send signal to the puppet
 	var pending = when.defer(),
 		ready = pending.promise;
 	if(!pid){
@@ -450,7 +263,7 @@ Master.prototype.debug = function debug(pid){
 		pending.resolve(true);
 	}
 	else{
-		var debugging = _this.slaves[pid];
+		var debugging = _this.puppets[pid];
 		_this.logger.info('[debug] going to debug live:%d status:%s', pid, debugging !== null);
 		ready = debugging.debug();
 
@@ -531,7 +344,7 @@ Master.prototype.fork = function(options, env){
 	var _this = this,
 		forked = cluster.fork(env);
 
-	_this.slaves[forked.process.pid] = new Slave(_this, forked, options, env);
+	_this.puppets[forked.process.pid] = new Puppet(_this, forked, options, env);
 
 	return forked;
 };
@@ -543,7 +356,7 @@ Master.prototype.whenStop = function whenStop(){
 		threshold = Date.now() + _this.stopTimeout,
 		gracefully = function(){
 
-			if(_.keys(_this.slaves).length === 0){//wait till each live slaves to be disconnected, then exit
+			if(_.keys(_this.puppets).length === 0){//wait till each live puppets to be disconnected, then exit
 				process.exit(0);
 			}
 
@@ -555,9 +368,9 @@ Master.prototype.whenStop = function whenStop(){
 			}
 		};//check every second
 
-	_.each(_this.slaves, function(slave){
+	_.each(_this.puppets, function(puppet){
 
-		slave.disconnect();
+		puppet.disconnect();
 	});
 
 	gracefully();//shutdown gracefully
@@ -565,9 +378,9 @@ Master.prototype.whenStop = function whenStop(){
 
 Master.prototype.whenExit = function whenExit(){
 
-	_.each(_this.slaves, function(slave){
+	_.each(_this.puppets, function(puppet){
 
-		slave.disconnect();
+		puppet.disconnect();
 	});
 
 	process.exit(-1);//shutdown bruteforcely
